refactor(server): extract physics, reducers, and tables into separate modules
Split monolithic lib.rs into domain-specific modules to adhere to Single Responsibility Principle, Modularity & Separation of Concerns, and Open/Closed Principle as per architectural review. Pure math/physics logic was moved to physics.rs, schema to tables.rs, and database logic to reducers.rs.
